+++ src/spectrogram/spectrogram_nonlinear.h
@@ -45,7 +45,7 @@ class SpectrogramNonlinear : public AlgorithmImplementation<SpectrogramConfigura
   private:
     void inline processAlgorithm(Input input, Output output)
     {
-        for (auto nFrame = 0; nFrame < Configuration::getNFrames(input.size(), C.bufferSize); nFrame++)
+        for (auto nFrame = 0; nFrame < Configuration::getNFrames(static_cast<int>(input.size()), C.bufferSize); nFrame++)
         {
             frame = input.segment(nFrame * C.bufferSize, C.bufferSize);
